vf128: canonicalize NaN values for non-zero mantissa and Inf exponent
- change NaN detection to simply require non-zero mantissa instead of
  copying the top 4-bits of the mantissa. Encode NaNs with the msb of
  the inline mantissa set i.e. `0x8`.

The prior logic for NaN values captured the top 4 bits of the mantissa.
That approach was sufficient for x86 quiet and signaling NaNs, but the
canonical rules for NaNs are that non-zero mantissa of the Inf exponent
are NaN values, so the prior logic would have erroneously encoded valid
NaN values as Inf because the top 4 bits were clear.

The revised logic encodes NaN by setting the msb of the inline mantissa.
The decode logic is split out but remains unchanged and will restore NaN
with the msb of the fraction set. The host should honor any non-zero
mantissa of the Inf exponent as a NaN value so this should be portable.

There is no support to indicate a signaling NaN.

References: #1