vf128: experimental normalized value compression support
implement an exponent compression scheme for succinct coding of values
between -0.99999.. to 0.99999.. excluding +/-0. the new scheme replaces
the current integer coding that uses the zero exponent code with a new
scheme that prepends exponent as a unary code prefix in the mantissa
adding only several bits of information versus an entire extra byte.

- most f32 values between -0.99999.. to 0.99999 fit 4 bytes or less
- most f64 values between -0.99999.. to 0.99999 fit 8 bytes or less

See #2